frontend/src/app/guards/auth/auth.guard.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
import { AuthService } from '../../services/auth/auth.service';

@Injectable({
  providedIn: 'root'
})
export class AuthGuard implements CanActivate {

  constructor(private authService: AuthService, private router: Router) {}

  canActivate(
    route: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ): boolean {

    const currentUser = localStorage.getItem('userId');
    const role = localStorage.getItem('role')

    if (!currentUser || !role) {
      this.router.navigate(['/login']);
      return false;
    }
    return true;
  }
}
